+++ src/utils.ts
@@ -6,10 +6,6 @@ import { IOptions } from './options';
 
 const readFile = util.promisify(fs.readFile);
 
-export const pathOptions = {
-    separator: '/'
-}
-
 export function first<T>(array: T[]): T {
     return array[0];
 }
